reorganized nodeDefs components folder structure
