+++ webapp/appModules/surveyForm/nodeDefs/components/nodeDeleteButton.js
